Fix: 예외 처리 통합 및 UserExceptionHandler 삭제
- GlobalExceptionHandler에서 예외 처리 로직을 통합하여 기존 UserExceptionHandler를 제거
- ErrorCode에서 불필요한 주석 제거
