+++ src/main/java/run/bemin/api/general/exception/GlobalExceptionHandler.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.NoHandlerFoundException;
+import run.bemin.api.user.exception.UserException;
 
 @RestControllerAdvice
 @Slf4j
@@ -116,5 +117,16 @@ public class GlobalExceptionHandler {
     final ErrorResponse response = ErrorResponse.of(ErrorCode.NOT_FOUND);
     return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
   }
+
+  /**
+   * 사용자 정의 예외(UserException) 발생할 경우
+   */
+  @ExceptionHandler(UserException.class)
+  protected ResponseEntity<ErrorResponse> handleUserException(UserException e) {
+    log.error("handleUserException", e);
+    final ErrorResponse response = ErrorResponse.of(e.getErrorCode());
+    return new ResponseEntity<>(response, HttpStatus.valueOf(e.getErrorCode().getStatus()));
+  }
+
 
 }
